OCR_RAG/requirements.txt
openai
streamlit
numpy
pandas
requests
pdf2image
faiss-cpu
python-dotenv
